added lz4 sandbox to ctest
