+++ src/cpp/CTestLists.txt
@@ -10,6 +10,7 @@ add_test(NAME diff_scheme_impl COMMAND test_diff_scheme_impl)
 #REDO tests with pipeline C API: add_test(NAME background_schemes COMMAND test_background_schemes)
 add_test(NAME background_scheme_impl COMMAND test_background_scheme_impl)
 add_test(NAME lz4_scheme_impl COMMAND test_lz4_scheme_impl)
+add_test(NAME lz4_sandbox COMMAND test_lz4_sandbox)
 add_test(NAME string_parsers_impl COMMAND test_string_parsers_impl)
 
 #REDO tests with pipeline C API: add_test(NAME bitswap_schemes COMMAND test_bitswap_schemes)
